requirements.txt
google-api-python-client>=2.100.0
snscrape>=0.6.0
feedparser>=6.0.0
praw>=7.7.0
python-telegram-bot==13.15
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.30.0
pytz>=2023.0
schedule>=1.2.0
